Moves files to match current project structure conventions #19
