+++ src/main/java/edu/oakland/ideas/model/Approver.java
@@ -1,4 +1,4 @@
-package edu.oakland.ideas.v1.model;
+package edu.oakland.ideas.model;
 
 import lombok.Data;
