+++ src/main/java/edu/oakland/ideas/service/IdeaDB.java
@@ -1,7 +1,7 @@
-package edu.oakland.ideas.v1.service;
+package edu.oakland.ideas.service;
 
-import edu.oakland.ideas.v1.model.*;
-import edu.oakland.ideas.v1.service.*;
+import edu.oakland.ideas.model.*;
+import edu.oakland.ideas.service.*;
 
 import java.sql.Timestamp;
 import java.util.HashMap;
